starting migration to Django 2.0
