+++ gibbs/forms.py
@@ -1,5 +1,5 @@
 from django import forms
-from gibbs import constants
+from util import constants
 import haystack.forms
 
 
@@ -149,16 +149,3 @@ class CompoundForm(BaseReactionForm):
         lambda self: self._GetWithDefault('submit', 'Update'))
 
 
-class BuildPathwayModelForm(forms.Form):
-    pathway_file = forms.FileField(required=True)
-    min_c = forms.FloatField(required=False)
-    max_c = forms.FloatField(required=False)
-    pH = forms.FloatField(required=False)
-    ionic_strength = forms.FloatField(required=False)
-    conc_units = forms.CharField(required=True)
-
-
-class AnalyzePathwayModelForm(forms.Form):
-    pathway_file = forms.FileField(required=True)
-    pH = forms.FloatField(required=False)
-    ionic_strength = forms.FloatField(required=False)
