+++ pathway/test/EMP_glycolysis_full_SBtab.tsv
