+++ pathway/test/EMP_glycolysis_simple.csv
