+++ requirements.txt
@@ -1,17 +1,17 @@
-django==1.10.4
-django-extensions==1.7.5
-django-haystack==2.5.1
-matplotlib==2.0.2
-MySQL-python==1.2.5
-nltk==3.2.1
-numpy==1.11.3
+django==2.0
+django-extensions==1.9.7
+django-haystack==2.6.1
+matplotlib==2.1.0
+mysqlclient==1.3.12
+nltk==3.2.5
+numpy==1.13.3
 pulp==1.6.8
-pyparsing==2.1.10
-scipy==0.18.1
-pandas==0.19.1
+pyparsing==2.2.0
+scipy==1.0.0
+pandas==0.21.0
 pysolr==3.6.0
-seaborn==0.7.1
+seaborn==0.8.1
 solr==0.4
-tablib==0.11.3
-xlrd==1.0.0
+tablib==0.12.1
+xlrd==1.1.0
