+++ util/SBtab/misc.py
@@ -63,7 +63,6 @@ def csv2xls(sbtab_file,delimiter):
        Delimiter of the column content; usually comma, tab, or semicolon.
     '''
     import xlwt
-    import tempfile
 
     book  = xlwt.Workbook()
     sheet = book.add_sheet('Sheet 1')
@@ -206,11 +205,11 @@ def parseReactionTable(sbtab_file,file_name,export=False):
     import SBtab
 
     if sbtab_file.table_type != 'Reaction':
-        print 'The given TableType \"%s\" cannot be parsed. The TableType \"Reaction\" is required.'%sbtab_file.table_type
+        print('The given TableType \"%s\" cannot be parsed. The TableType \"Reaction\" is required.'%sbtab_file.table_type)
         return False
 
     if not '!ReactionFormula' in sbtab_file.columns:
-        print 'The given provided SBtab file misses the column \"ReactionFormula\".'
+        print('The given provided SBtab file misses the column \"ReactionFormula\".')
         return False
     else:
         for i,c in enumerate(sbtab_file.columns):
